Create MinimumElements
An implementation of a function minimumElements that takes an array of integers num and an integer x as input. The goal of the function is to determine the minimum number of elements from the array num that need to be selected in order to form a sum equal to or greater than the target value x.